refactor: use value selector for hold-to-unpause duration
Replaced PreferenceTimeStepper with PreferenceValueSelector for the "Hold to Unpause Duration" setting. This provides a cleaner UX with predefined options (0.5s, 1s, 2s, 3s, 5s) instead of a continuous range, matching the pattern used for Pause Timeout.

Changes:
- Updated PauseSettingsScreenModel to use Int instead of Long for hold-to-unpause
- Replaced PreferenceTimeStepper with PreferenceValueSelector
- Defined five duration options: 0.5s, 1s, 2s, 3s, 5s
- Added button and display formatters for user-friendly labels
- Removed unused PreferenceTimeStepper import

Both pause settings now use consistent value selector UI.

Closes #1989

+++ app/src/main/java/com/enaboapps/switchify/screens/settings/pause/models/PauseSettingsScreenModel.kt
@@ -9,14 +9,14 @@ import com.enaboapps.switchify.backend.preferences.PreferenceManager
 class PauseSettingsScreenModel(context: Context) : ViewModel() {
     private val preferenceManager = PreferenceManager(context)
 
-    private val _holdToUnpauseDuration = MutableLiveData<Long>().apply {
+    private val _holdToUnpauseDuration = MutableLiveData<Int>().apply {
         value =
             preferenceManager.getLongValue(
                 PreferenceManager.Keys.PREFERENCE_KEY_HOLD_TO_UNPAUSE_DURATION,
                 2000L
-            )
+            ).toInt()
     }
-    val holdToUnpauseDuration: LiveData<Long> = _holdToUnpauseDuration
+    val holdToUnpauseDuration: LiveData<Int> = _holdToUnpauseDuration
 
     private val _pauseTimeout = MutableLiveData<Int>().apply {
         value =
@@ -27,10 +27,10 @@ class PauseSettingsScreenModel(context: Context) : ViewModel() {
     }
     val pauseTimeout: LiveData<Int> = _pauseTimeout
 
-    fun setHoldToUnpauseDuration(value: Long) {
+    fun setHoldToUnpauseDuration(value: Int) {
         preferenceManager.setLongValue(
             PreferenceManager.Keys.PREFERENCE_KEY_HOLD_TO_UNPAUSE_DURATION,
-            value
+            value.toLong()
         )
         _holdToUnpauseDuration.postValue(value)
     }
